fix: correct logic to test for scope
`(match-string scope-index)` does not return a 'nil even when there is no match
for the scope. This is probably because of the fact that the scope is an
optional match. In this commit, we fix this logic by asserting that the
existence of a scope is implied when the position for the end of the match is
larger in value than the start. This means that it cannot be a zero-width
character.